+++ entrypoint/__init__.py
@@ -28,8 +28,8 @@ __url__ = "https://github.com/nekitdev/entrypoint"
 __title__ = "entrypoint"
 __author__ = "nekitdev"
 __license__ = "MIT"
-__version__ = "1.3.0"
+__version__ = "1.4.0"
 
-from entrypoint.core import MAIN, EntryPoint, entrypoint, is_main
+from entrypoint.core import MAIN, EntryPoint, Main, entrypoint, is_main
 
-__all__ = ("MAIN", "EntryPoint", "entrypoint", "is_main")
+__all__ = ("MAIN", "Main", "EntryPoint", "entrypoint", "is_main")